test(integration): Split integration tests into separate files.
